requirements.txt
boto3==1.34.96
openai==1.30.1
python-dotenv==1.0.1
selenium==4.21.0
